+++ src/ts/Ads/Managers/LegacyCampaignManager.ts
@@ -3,7 +3,7 @@ import { AssetManager } from 'Ads/Managers/AssetManager';
 import { RefreshManager } from 'Ads/Managers/RefreshManager';
 import { SessionManager } from 'Ads/Managers/SessionManager';
 import { AdsConfiguration } from 'Ads/Models/AdsConfiguration';
-import { AuctionResponse, IRawAuctionResponse, IRawAuctionV5Response, AuctionStatusCode } from 'Ads/Models/AuctionResponse';
+import { AuctionResponse, IRawAuctionResponse, IRawAuctionV5Response, AuctionStatusCode, IRawAuctionV6Response } from 'Ads/Models/AuctionResponse';
 import { Campaign, ICampaignTrackingUrls } from 'Ads/Models/Campaign';
 import { Placement } from 'Ads/Models/Placement';
 import { Session } from 'Ads/Models/Session';
@@ -44,13 +44,12 @@ import { CampaignContentTypes } from 'Ads/Utilities/CampaignContentTypes';
 import { ProgrammaticVastParser } from 'VAST/Parsers/ProgrammaticVastParser';
 import { TrackingIdentifierFilter } from 'Ads/Utilities/TrackingIdentifierFilter';
 import { VastCampaign } from 'VAST/Models/VastCampaign';
-import { SDKMetrics, LoadMetric, GeneralTimingMetric } from 'Ads/Utilities/SDKMetrics';
+import { SDKMetrics, LoadMetric, GeneralTimingMetric, MiscellaneousMetric } from 'Ads/Utilities/SDKMetrics';
 import { PrivacySDK } from 'Privacy/PrivacySDK';
 import { PARTNER_NAME, OM_JS_VERSION } from 'Ads/Views/OpenMeasurement/OpenMeasurement';
 import { UserPrivacyManager } from 'Ads/Managers/UserPrivacyManager';
 import { MediationLoadTrackingManager, MediationExperimentType } from 'Ads/Managers/MediationLoadTrackingManager';
 import { createMeasurementsInstance, ITimeMeasurements } from 'Core/Utilities/TimeMeasurements';
-import { SdkDetectionInfo } from 'Core/Models/SdkDetectionInfo';
 import { CampaignManager } from 'Ads/Managers/CampaignManager';
 import { XHRequest } from 'Core/Utilities/XHRequest';
 
@@ -75,14 +74,12 @@ export class LegacyCampaignManager extends CampaignManager {
     private _metaDataManager: MetaDataManager;
     private _request: RequestManager;
     private _deviceInfo: DeviceInfo;
-    private _previousPlacementId: string | undefined;
     private _lastAuctionId: string | undefined;
     private _deviceFreeSpace: number;
     private _auctionProtocol: AuctionProtocol;
     private _isLoadEnabled: boolean = false;
     private _userPrivacyManager: UserPrivacyManager;
     private _mediationLoadTracking: MediationLoadTrackingManager | undefined;
-    private _sdkDetectionInfo: SdkDetectionInfo;
 
     constructor(platform: Platform, core: ICore, coreConfig: CoreConfiguration, adsConfig: AdsConfiguration, assetManager: AssetManager, sessionManager: SessionManager, adMobSignalFactory: AdMobSignalFactory, request: RequestManager, clientInfo: ClientInfo, deviceInfo: DeviceInfo, metaDataManager: MetaDataManager, cacheBookkeeping: CacheBookkeepingManager, contentTypeHandlerManager: ContentTypeHandlerManager, privacySDK: PrivacySDK, userPrivacyManager: UserPrivacyManager, mediationLoadTracking?: MediationLoadTrackingManager | undefined) {
         super();
@@ -105,7 +102,6 @@ export class LegacyCampaignManager extends CampaignManager {
         this._privacy = privacySDK;
         this._userPrivacyManager = userPrivacyManager;
         this._mediationLoadTracking = mediationLoadTracking;
-        this._sdkDetectionInfo = core.SdkDetectionInfo;
     }
 
     public request(nofillRetry?: boolean): Promise<INativeResponse | void> {
@@ -115,6 +111,15 @@ export class LegacyCampaignManager extends CampaignManager {
         // prevent having more then one ad request in flight
         if (this._requesting) {
             return Promise.resolve();
+        }
+
+        SDKMetrics.reportMetricEventWithTags(MiscellaneousMetric.AuctionRequestCreated, {
+            'wel': 'false',
+            'iar': `${GameSessionCounters.getCurrentCounters().adRequests === 1}`
+        });
+
+        if (this._mediationLoadTracking) {
+            this._mediationLoadTracking.reportAuctionRequestStarted();
         }
 
         GameSessionCounters.addAdRequest();
@@ -132,14 +137,14 @@ export class LegacyCampaignManager extends CampaignManager {
         const requestTimestamp: number = Date.now();
 
         return Promise.all<string[], number | undefined, number>([
-            this.getFullyCachedCampaigns(),
-            this.getVersionCode(),
+            CampaignManager.getFullyCachedCampaigns(this._core),
+            CampaignManager.getVersionCode(this._platform, this._core, this._clientInfo),
             this._deviceInfo.getFreeSpace()
         ]).then(([fullyCachedCampaignIds, versionCode, freeSpace]) => {
             this._deviceFreeSpace = freeSpace;
             return Promise.all<string, unknown>([
                 CampaignManager.createRequestUrl(this.getBaseUrl(), this._platform, this._clientInfo, this._deviceInfo, this._coreConfig, this._lastAuctionId, nofillRetry),
-                CampaignManager.createRequestBody(this._clientInfo, this._coreConfig, this._deviceInfo, this._userPrivacyManager, this._sessionManager, this._privacy, countersForOperativeEvents, fullyCachedCampaignIds, versionCode, this._adMobSignalFactory, freeSpace, this._metaDataManager, this._adsConfig, this._isLoadEnabled, this.getPreviousPlacementId(), requestPrivacy, legacyRequestPrivacy, nofillRetry, this._sdkDetectionInfo)
+                CampaignManager.createRequestBody(this._clientInfo, this._coreConfig, this._deviceInfo, this._userPrivacyManager, this._sessionManager, this._privacy, countersForOperativeEvents, fullyCachedCampaignIds, versionCode, this._adMobSignalFactory, freeSpace, this._metaDataManager, this._adsConfig, this._isLoadEnabled, this.getPreviousPlacementId(), requestPrivacy, legacyRequestPrivacy, nofillRetry)
             ]);
         }).then(([requestUrl, requestBody]) => {
             measurement = createMeasurementsInstance(GeneralTimingMetric.AuctionRequest, {
@@ -161,11 +166,28 @@ export class LegacyCampaignManager extends CampaignManager {
                 return CampaignManager.onlyRequest(this._request, requestUrl, requestBody);
             }
         }).catch((error: unknown) => {
-            if (this._mediationLoadTracking && performance && performance.now) {
-                this._mediationLoadTracking.reportAuctionRequest(this.getTime() - requestStartTime, false);
+            let reason: string = 'unknown';
+            if (error instanceof RequestError) {
+                if (error.nativeResponse) {
+                    reason = error.nativeResponse.responseCode.toString();
+                } else {
+                    reason = 'request';
+                }
             }
+            if (this._mediationLoadTracking && performance && performance.now) {
+                this._mediationLoadTracking.reportAuctionRequest(this.getTime() - requestStartTime, false, reason);
+            }
+            SDKMetrics.reportMetricEventWithTags(MiscellaneousMetric.AuctionRequestFailed, {
+                'wel': 'false',
+                'iar': `${GameSessionCounters.getCurrentCounters().adRequests === 1}`,
+                'rsn': reason
+            });
             throw error;
         }).then(response => {
+            SDKMetrics.reportMetricEventWithTags(MiscellaneousMetric.AuctionRequestOk, {
+                'wel': 'false',
+                'iar': `${GameSessionCounters.getCurrentCounters().adRequests === 1}`
+            });
             measurement.measure('auction_response');
             const cachingTime = this.getTime();
             if (this._mediationLoadTracking && performance && performance.now) {
@@ -174,22 +196,28 @@ export class LegacyCampaignManager extends CampaignManager {
             if (response) {
                 this.setSDKSignalValues(requestTimestamp);
 
-                if (this._auctionProtocol === AuctionProtocol.V5) {
-                    return this.parseAuctionV5Campaigns(response, countersForOperativeEvents, requestPrivacy, legacyRequestPrivacy).then(() => {
-                        if (this._mediationLoadTracking && performance && performance.now) {
-                            this._mediationLoadTracking.reportingAdCaching(this.getTime() - cachingTime, true);
-                        }
-                    }).catch((e) => {
-                        if (this._mediationLoadTracking && performance && performance.now) {
-                            this._mediationLoadTracking.reportingAdCaching(this.getTime() - cachingTime, false);
-                        }
-                        this.handleGeneralError(e, 'parse_auction_v5_campaigns_error');
-                    });
-                } else {
-                    return this.parseCampaigns(response, countersForOperativeEvents, requestPrivacy, legacyRequestPrivacy).catch((e) => {
-                        this.handleGeneralError(e, 'parse_campaigns_error');
-                    });
+                let parseResponse: Promise<void | void[]>;
+                switch (this._auctionProtocol) {
+                    case AuctionProtocol.V6:
+                        parseResponse = this.parseAuctionV6Campaigns(response, countersForOperativeEvents, requestPrivacy, legacyRequestPrivacy);
+                    case AuctionProtocol.V5:
+                        parseResponse = this.parseAuctionV5Campaigns(response, countersForOperativeEvents, requestPrivacy, legacyRequestPrivacy);
+                        break;
+                    case AuctionProtocol.V4:
+                    default:
+                        parseResponse = this.parseCampaigns(response, countersForOperativeEvents, requestPrivacy, legacyRequestPrivacy);
                 }
+
+                return parseResponse.then(() => {
+                    if (this._mediationLoadTracking && performance && performance.now) {
+                        this._mediationLoadTracking.reportingAdCaching(this.getTime() - cachingTime, true);
+                    }
+                }).catch((e) => {
+                    if (this._mediationLoadTracking && performance && performance.now) {
+                        this._mediationLoadTracking.reportingAdCaching(this.getTime() - cachingTime, false);
+                    }
+                    this.handleGeneralError(e, 'parse_auction_campaigns_error');
+                });
             }
             throw new WebViewError('Empty campaign response', 'CampaignRequestError');
         }).then(() => {
@@ -207,6 +235,15 @@ export class LegacyCampaignManager extends CampaignManager {
     }
 
     public loadCampaign(placement: Placement): Promise<ILoadedCampaign | undefined> {
+        SDKMetrics.reportMetricEventWithTags(MiscellaneousMetric.AuctionRequestCreated, {
+            'wel': 'true',
+            'iar': `${GameSessionCounters.getCurrentCounters().adRequests === 1}`
+        });
+
+        if (this._mediationLoadTracking) {
+            this._mediationLoadTracking.reportAuctionRequestStarted();
+        }
+
         const requestStartTime = this.getTime();
         let cachingTime: number;
         this._isLoadEnabled = true;
@@ -222,14 +259,15 @@ export class LegacyCampaignManager extends CampaignManager {
         const legacyRequestPrivacy = RequestPrivacyFactory.createLegacy(this._privacy);
 
         return Promise.all<string[], number | undefined, number>([
-            this.getFullyCachedCampaigns(),
-            this.getVersionCode(),
+            CampaignManager.getFullyCachedCampaigns(this._core),
+            CampaignManager.getVersionCode(this._platform, this._core, this._clientInfo),
             this._deviceInfo.getFreeSpace()
         ]).then(([fullyCachedCampaignIds, versionCode, freeSpace]) => {
             this._deviceFreeSpace = freeSpace;
             return Promise.all<string, unknown>([
-                CampaignManager.createRequestUrl(this.getBaseUrl(), this._platform, this._clientInfo, this._deviceInfo, this._coreConfig, this._lastAuctionId, false),
-                CampaignManager.createRequestBody(this._clientInfo, this._coreConfig, this._deviceInfo, this._userPrivacyManager, this._sessionManager, this._privacy, countersForOperativeEvents, fullyCachedCampaignIds, versionCode, this._adMobSignalFactory, freeSpace, this._metaDataManager, this._adsConfig, this._isLoadEnabled, this.getPreviousPlacementId(), requestPrivacy, legacyRequestPrivacy, false, this._sdkDetectionInfo, placement)
+                // TODO: Utilize this.getBaseUrl() after V6 is supported for load
+                CampaignManager.createRequestUrl(this.constructBaseUrl(CampaignManager.AuctionV5BaseUrl), this._platform, this._clientInfo, this._deviceInfo, this._coreConfig, this._lastAuctionId, false),
+                CampaignManager.createRequestBody(this._clientInfo, this._coreConfig, this._deviceInfo, this._userPrivacyManager, this._sessionManager, this._privacy, countersForOperativeEvents, fullyCachedCampaignIds, versionCode, this._adMobSignalFactory, freeSpace, this._metaDataManager, this._adsConfig, this._isLoadEnabled, this.getPreviousPlacementId(), requestPrivacy, legacyRequestPrivacy, false, placement)
             ]);
         }).then(([requestUrl, requestBody]) => {
             this._core.Sdk.logInfo('Loading placement ' + placement.getId() + ' from ' + requestUrl);
@@ -244,12 +282,29 @@ export class LegacyCampaignManager extends CampaignManager {
                     retryWithConnectionEvents: false,
                     timeout: 10000
                 }).catch((error: unknown) => {
-                    if (this._mediationLoadTracking && performance && performance.now) {
-                        this._mediationLoadTracking.reportAuctionRequest(this.getTime() - requestStartTime, false);
+                    let reason: string = 'unknown';
+                    if (error instanceof RequestError) {
+                        if (error.nativeResponse) {
+                            reason = error.nativeResponse.responseCode.toString();
+                        } else {
+                            reason = 'request';
+                        }
                     }
+                    if (this._mediationLoadTracking && performance && performance.now) {
+                        this._mediationLoadTracking.reportAuctionRequest(this.getTime() - requestStartTime, false, reason);
+                    }
+                    SDKMetrics.reportMetricEventWithTags(MiscellaneousMetric.AuctionRequestFailed, {
+                        'wel': 'true',
+                        'iar': `${GameSessionCounters.getCurrentCounters().adRequests === 1}`,
+                        'rsn': reason
+                    });
                     throw error;
                 });
             }).then(response => {
+                SDKMetrics.reportMetricEventWithTags(MiscellaneousMetric.AuctionRequestOk, {
+                    'wel': 'true',
+                    'iar': `${GameSessionCounters.getCurrentCounters().adRequests === 1}`
+                });
                 cachingTime = this.getTime();
                 if (this._mediationLoadTracking && performance && performance.now) {
                     this._mediationLoadTracking.reportAuctionRequest(this.getTime() - requestStartTime, true);
@@ -274,22 +329,6 @@ export class LegacyCampaignManager extends CampaignManager {
                 SDKMetrics.reportMetricEvent(LoadMetric.LoadEnabledNoFill);
                 return undefined;
             });
-        });
-    }
-
-    public setPreviousPlacementId(id: string | undefined) {
-        this._previousPlacementId = id;
-    }
-
-    public getPreviousPlacementId(): string | undefined {
-        return this._previousPlacementId;
-    }
-
-    public getFullyCachedCampaigns(): Promise<string[]> {
-        return this._core.Storage.getKeys(StorageType.PRIVATE, 'cache.campaigns', false).then((campaignKeys) => {
-            return campaignKeys;
-        }).catch(() => {
-            return [];
         });
     }
 
@@ -399,6 +438,17 @@ export class LegacyCampaignManager extends CampaignManager {
         } else {
             throw new Error('No placements found');
         }
+    }
+
+    // TODO: Update with implementation
+    private parseAuctionV6Campaigns(response: INativeResponse, gameSessionCounters: IGameSessionCounters, requestPrivacy?: IRequestPrivacy, legacyRequestPrivacy?: ILegacyRequestPrivacy): Promise<void> {
+        let json;
+        try {
+            json = JsonParser.parse<IRawAuctionV6Response>(response.response);
+        } catch (e) {
+            return Promise.reject(new Error('Could not parse campaign JSON: ' + e.message));
+        }
+        return Promise.resolve();
     }
 
     private parseAuctionV5Campaigns(response: INativeResponse, gameSessionCounters: IGameSessionCounters, requestPrivacy?: IRequestPrivacy, legacyRequestPrivacy?: ILegacyRequestPrivacy): Promise<void[]> {
@@ -547,6 +597,7 @@ export class LegacyCampaignManager extends CampaignManager {
     private parseLoadedCampaign(response: INativeResponse, placement: Placement, gameSessionCounters: IGameSessionCounters, deviceFreeSpace: number, requestPrivacy?: IRequestPrivacy, legacyRequestPrivacy?: ILegacyRequestPrivacy): Promise<ILoadedCampaign | undefined> {
         let json;
         try {
+            // TODO: Transition Load to use Auction V6
             json = JsonParser.parse<IRawAuctionV5Response>(response.response);
         } catch (e) {
             Diagnostics.trigger('load_campaign_failed_to_parse', {});
@@ -773,41 +824,26 @@ export class LegacyCampaignManager extends CampaignManager {
 
     private getBaseUrl(): string {
         if (this._coreConfig.getTestMode()) {
-            return [
-                CampaignManager.TestModeUrl,
-                this._clientInfo.getGameId(),
-                'requests'
-            ].join('/');
+            return this.constructBaseUrl(CampaignManager.TestModeUrl);
         }
-        if (this._auctionProtocol === AuctionProtocol.V5) {
-            return [
-                CampaignManager.AuctionV5BaseUrl,
-                this._clientInfo.getGameId(),
-                'requests'
-            ].join('/');
-        } else {
-            return [
-                CampaignManager.BaseUrl,
-                this._clientInfo.getGameId(),
-                'requests'
-            ].join('/');
+
+        switch (this._auctionProtocol) {
+            case AuctionProtocol.V6:
+                return this.constructBaseUrl(CampaignManager.AuctionV6BaseUrl);
+            case AuctionProtocol.V5:
+                return this.constructBaseUrl(CampaignManager.AuctionV5BaseUrl);
+            case AuctionProtocol.V4:
+            default:
+                return this.constructBaseUrl(CampaignManager.BaseUrl);
         }
     }
 
-    private getVersionCode(): Promise<number | undefined> {
-        if (this._platform === Platform.ANDROID) {
-            return this._core.DeviceInfo.Android!.getPackageInfo(this._clientInfo.getApplicationName()).then(packageInfo => {
-                if (packageInfo.versionCode) {
-                    return packageInfo.versionCode;
-                } else {
-                    return undefined;
-                }
-            }).catch(() => {
-                return undefined;
-            });
-        } else {
-            return Promise.resolve(undefined);
-        }
+    private constructBaseUrl(baseUri: string): string {
+        return [
+            baseUri,
+            this._clientInfo.getGameId(),
+            'requests'
+        ].join('/');
     }
 
     private setSDKSignalValues(requestTimestamp: number): void {