+++ test/Unit/Ads/UserPrivacyManagerTest.ts
@@ -651,7 +651,8 @@ describe('UserPrivacyManagerTest', () => {
                 bundleId: testBundleId,
                 permissions: permissions,
                 legalFramework: 'none',
-                agreedOverAgeLimit: 'missing' };
+                agreedOverAgeLimit: 'missing'
+            };
 
             return privacyManager.updateUserPrivacy(permissions, source, action, layout).then(() => {
                 const sentKafkaObject = httpKafkaSpy.firstCall.args[2];